@@ -1,64 +1,77 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 using UndertaleModLib.Models;
 
 namespace UndertaleModLib.Decompiler
 {
     public static class Disassembler
     {
-        public static string GenerateLocalVarDefinitions(this UndertaleCode code, IList<UndertaleVariable> vars, UndertaleCodeLocals locals)
+        private static void AppendLocalVarDefinitionsToStringBuilder(StringBuilder sb, UndertaleCode code, IList<UndertaleVariable> vars, UndertaleCodeLocals locals)
         {
-            if (code.WeirdLocalFlag)
-                return "";
-            if (locals == null)
-                return "; Missing code locals- possibly due to unsupported bytecode version or brand new code entry.\n";
-
-            StringBuilder sb = new StringBuilder();
-
-            var referenced = code.FindReferencedLocalVars();
-            if (locals.Name != code.Name)
-                throw new Exception("Name of the locals block does not match name of the code block");
-            foreach (var arg in locals.Locals)
+            if (code.WeirdLocalFlag && locals is null)
             {
-                sb.Append(".localvar " + arg.Index + " " + arg.Name.Content);
-                var refvar = referenced.Where((x) => x.Name == arg.Name && x.VarID == arg.Index).FirstOrDefault();
-                if (refvar != null)
-                {
-                    sb.Append(" " + vars.IndexOf(refvar));
-                }
-                sb.Append('\n');
+                return;
             }
 
-            return sb.ToString();
+            if (locals is null)
+            {
+                sb.AppendLine("; Missing code locals, possibly due to unsupported bytecode version or brand new code entry.");
+                return;
+            }
+            
+            var referenced = code.FindReferencedLocalVars();
+            if (locals.Name != code.Name)
+                throw new Exception("Name of the locals block does not match name of the code block!");
+            foreach (var arg in locals.Locals)
+            {
+                sb.Append(".localvar ");
+                sb.Append(arg.Index);
+                sb.Append(' ');
+                sb.Append(arg.Name.Content);
+                var refVar = referenced.FirstOrDefault(x => x.Name == arg.Name && x.VarID == arg.Index);
+                if (refVar is not null)
+                {
+                    sb.Append(' ');
+                    sb.Append(vars.IndexOf(refVar));
+                }
+                sb.AppendLine();
+            }
         }
-
+        
         public static string Disassemble(this UndertaleCode code, IList<UndertaleVariable> vars, UndertaleCodeLocals locals)
         {
-            StringBuilder sb = new StringBuilder();
-            if (locals == null && !code.WeirdLocalFlag)
-                sb.Append("; WARNING: Missing code locals, possibly due to unsupported bytecode version or a brand new code entry.\n");
+            // This StringBuilder is shared with the ToString method of the code instructions.
+            // Experimentation has shown that 200 is a good enough starting value for it (now changed to 256).
+            // 300 seemed too high and 100 too low. This may change in the future.
+            StringBuilder sb = new(256);
+            if (locals is null && !code.WeirdLocalFlag)
+                sb.AppendLine("; WARNING: Missing code locals, possibly due to unsupported bytecode version or a brand new code entry.");
             else
-                sb.Append(code.GenerateLocalVarDefinitions(vars, locals));
+                AppendLocalVarDefinitionsToStringBuilder(sb, code, vars, locals);
 
-            Dictionary<uint, string> fragments = new Dictionary<uint, string>();
+            Dictionary<uint, string> fragments = new(code.ChildEntries.Count);
             foreach (var dup in code.ChildEntries)
-                fragments.Add(dup.Offset / 4, (dup.Name?.Content ?? "<null>") + $" (locals={dup.LocalsCount}, argc={dup.ArgumentsCount})");
+            {
+                fragments.Add(dup.Offset / 4, $"{(dup.Name?.Content ?? "<null>")} (locals={dup.LocalsCount}, argc={dup.ArgumentsCount})");
+            }
+
             List<uint> blocks = FindBlockAddresses(code);
 
+            uint address = 0;
             foreach (var inst in code.Instructions)
             {
                 bool doNewline = true;
-                if (fragments.TryGetValue(inst.Address, out string entry))
+                if (fragments.TryGetValue(address, out string entry))
                 {
                     sb.AppendLine();
                     sb.AppendLine($"> {entry}");
                     doNewline = false;
                 }
 
-                int ind = blocks.IndexOf(inst.Address);
+                int ind = blocks.IndexOf(address);
                 if (ind != -1)
                 {
                     if (doNewline)
@@ -66,7 +79,10 @@ namespace UndertaleModLib.Decompiler
                     sb.AppendLine($":[{ind}]");
                 }
 
-                sb.AppendLine(inst.ToString(code, blocks));
+                inst.ToString(sb, code, address, blocks);
+                sb.AppendLine();
+
+                address += inst.CalculateInstructionSize();
             }
 
             sb.AppendLine();
@@ -82,6 +98,7 @@ namespace UndertaleModLib.Decompiler
             if (code.Instructions.Count != 0)
                 addresses.Add(0);
 
+            uint address = 0;
             foreach (var inst in code.Instructions)
             {
                 switch (inst.Kind)
@@ -90,18 +107,19 @@ namespace UndertaleModLib.Decompiler
                     case UndertaleInstruction.Opcode.Bf:
                     case UndertaleInstruction.Opcode.Bt:
                     case UndertaleInstruction.Opcode.PushEnv:
-                        addresses.Add(inst.Address + 1);
-                        addresses.Add((uint)(inst.Address + inst.JumpOffset));
+                        addresses.Add(address + 1);
+                        addresses.Add((uint)(address + inst.JumpOffset));
                         break;
                     case UndertaleInstruction.Opcode.PopEnv:
                         if (!inst.JumpOffsetPopenvExitMagic)
-                            addresses.Add((uint)(inst.Address + inst.JumpOffset));
+                            addresses.Add((uint)(address + inst.JumpOffset));
                         break;
                     case UndertaleInstruction.Opcode.Exit:
                     case UndertaleInstruction.Opcode.Ret:
-                        addresses.Add(inst.Address + 1);
+                        addresses.Add(address + 1);
                         break;
                 }
+                address += inst.CalculateInstructionSize();
             }
 
             List<uint> res = addresses.ToList();
