@@ -60,9 +60,10 @@ namespace UndertaleModLib
         private static ChunkT FindListChunk(UndertaleData data)
         {
             if (data.FORM.ChunksTypeDict.TryGetValue(typeof(ChunkT), out UndertaleChunk chunk))
-                return chunk as ChunkT;
-            else
-                return null;
+            {
+                return (ChunkT)chunk;
+            }
+            return null;
         }
 
         public int SerializeById(UndertaleWriter writer)
@@ -135,8 +136,7 @@ namespace UndertaleModLib
 
         public void Serialize(UndertaleWriter writer)
         {
-            writer.resourceIDRefsToWrite.Add(new ValueTuple<uint, UndertaleResourceRef>(writer.Position, this));
-            writer.Write((int)0);
+            writer.Write(SerializeById(writer));
         }
 
         public void Unserialize(UndertaleReader reader)
@@ -174,6 +174,14 @@ namespace UndertaleModLib
         /// </summary>
         public string Directory { get; set; } = null;
 
+        internal readonly record struct BytecodeInformation(uint InstructionCount, UndertaleCode RootEntry);
+
+        internal Dictionary<uint, BytecodeInformation> BytecodeAddresses;
+        internal ArrayPool<uint> ListPtrsPool = ArrayPool<uint>.Create(100000, 17);
+        internal string LastChunkName;
+        internal List<string> AllChunkNames;
+        internal bool Bytecode14OrLower = false;
+
         public UndertaleReader(Stream input,
                                WarningHandlerDelegate warningHandler = null, MessageHandlerDelegate messageHandler = null,
                                bool onlyGeneralInfo = false) : base(input)
@@ -206,11 +214,6 @@ namespace UndertaleModLib
             else
                 Debug.WriteLine(message);
         }
-
-        public string LastChunkName;
-        public List<string> AllChunkNames;
-        //public bool GMS2_3 = false;
-        public bool Bytecode14OrLower = false;
 
         public UndertaleChunk ReadUndertaleChunk()
         {
@@ -256,7 +259,7 @@ namespace UndertaleModLib
                     SwitchReaderType(false);
                 }
             }
-            utListPtrsPool = null;
+            ListPtrsPool = null;
 
             InitializePools(poolSize);
 
@@ -275,8 +278,24 @@ namespace UndertaleModLib
             if (!FilePath.EndsWith(".dat"))
             {
                 data.BuiltinList = new BuiltinList(data);
-                Decompiler.AssetTypeResolver.InitializeTypes(data);
+                Decompiler.GameSpecificResolver.Initialize(data);
                 UndertaleEmbeddedTexture.FindAllTextureInfo(data);
+            }
+
+            // Iterate over function names to see if 2023.11+ naming process was used (if necessary)
+            if (data.Functions is not null && data.IsVersionAtLeast(2023, 8) && !data.IsVersionAtLeast(2023, 11))
+            {
+                foreach (UndertaleFunction function in data.Functions)
+                {
+                    // If name starts with "gml_Script" and contains a @ character, it should be from 2023.11
+                    if (function.Name.Content is string functionName &&
+                        functionName.StartsWith("gml_Script_", StringComparison.Ordinal) && 
+                        functionName.Contains('@'))
+                    {
+                        data.SetGMS2Version(2023, 11);
+                        break;
+                    }
+                }
             }
 
             ProcessCountExc(poolSize);
@@ -307,16 +326,12 @@ namespace UndertaleModLib
         private readonly Dictionary<Type, Func<UndertaleReader, uint>> unserializeFuncDict = new();
         private readonly Dictionary<Type, uint> staticObjCountDict = new();
         private readonly Dictionary<Type, uint> staticObjSizeDict = new();
-        public HashSet<uint> GMS2BytecodeAddresses;
-        public int[] InstructionArraysLengths;
 
         private readonly BindingFlags publicStaticFlags
             = BindingFlags.Public | BindingFlags.Static | BindingFlags.FlattenHierarchy;
         private readonly Type[] readerArgType = { typeof(UndertaleReader) };
         private readonly Type delegateType = typeof(Func<UndertaleReader, uint>);
         private readonly Func<UndertaleReader, uint> blankCountFunc = new(_ => { return 0; });
-
-        public ArrayPool<uint> utListPtrsPool = ArrayPool<uint>.Create(100000, 17);
 
         private bool ProcessCountExc(uint poolSize = 0)
         {
@@ -344,6 +359,7 @@ namespace UndertaleModLib
             if (poolSize != 0 && poolSize != objectPool.Count)
             {
                 SubmitWarning("Warning - the estimated object pool size differs from the actual size.\n" +
+                             $"Estimated: {poolSize}, Actual: {objectPool.Count}\n" +
                               "Please report this on UndertaleModTool GitHub.");
             }
 
@@ -587,9 +603,23 @@ namespace UndertaleModLib
 
         public T ReadUndertaleObject<T>() where T : UndertaleObject, new()
         {
-            T obj = GetUndertaleObjectAtAddress<T>((uint)AbsPosition);
-            ReadUndertaleObject(obj);
-            return obj;
+            uint address = (uint)AbsPosition;
+
+            T result;
+            if (objectPool.TryGetValue(address, out UndertaleObject obj))
+            {
+                result = (T)obj;
+                unreadObjects.Remove(address);
+            }
+            else
+            {
+                result = new T();
+                objectPool.Add(address, result);
+                objectPoolRev.Add(result, address);
+            }
+
+            result.Unserialize(this);
+            return result;
         }
 
         public T ReadUndertaleObjectPointer<T>() where T : UndertaleObject, new()
@@ -638,7 +668,7 @@ namespace UndertaleModLib
                 if (length != expectedLength)
                 {
                     int diff = (int)expectedLength - (int)length;
-                    reader.SubmitMessage("WARNING: File specified length " + expectedLength + ", but read only " + length + " (" + diff + " padding?)");
+                    reader.SubmitWarning("WARNING: File specified length " + expectedLength + ", but read only " + length + " (" + diff + " padding?)");
                     if (diff > 0)
                         reader.Position += (uint)diff;
                     else
@@ -700,61 +730,37 @@ namespace UndertaleModLib
             undertaleData = data;
             Bytecode14OrLower = data?.GeneralInfo?.BytecodeVersion <= 14;
 
-            // Figure out the last chunk by iterating identically as it does when serializing
+            // Figure out the last chunk by iterating identically as it does when serializing,
+            // and generate the object index dictionaries for acceleration of "UndertaleResourceById.SerializeById()"
             foreach (var chunk in data.FORM.Chunks)
             {
                 LastChunkName = chunk.Key;
-            }
 
-            // Generate the object index dictionaries for acceleration of "UndertaleResourceById.SerializeById()"
-            var listChunks = data.FORM.Chunks.Values.Select(x => x as IUndertaleListChunk);
-            Parallel.ForEach(listChunks.Where(x => x is not null), (chunk) =>
-            {
-                chunk.GenerateIndexDict();
-            });
-            UndertaleIO.IsDictionaryCleared = false;
+                if (chunk.Value is IUndertaleListChunk listChunk)
+                {
+                    listChunk.GenerateIndexDict();
+                }
+            }
 
             Write(data.FORM);
         }
 
-        private Dictionary<UndertaleObject, uint> objectPool = new Dictionary<UndertaleObject, uint>();
-        private Dictionary<UndertaleObject, List<uint>> pendingWrites = new Dictionary<UndertaleObject, List<uint>>();
-        private Dictionary<UndertaleObject, List<uint>> pendingStringWrites = new Dictionary<UndertaleObject, List<uint>>();
-        private List<ValueTuple<uint, uint>> intsToWriteParallel = new List<ValueTuple<uint, uint>>();
-        public List<ValueTuple<uint, UndertaleResourceRef>> resourceIDRefsToWrite = new List<ValueTuple<uint, UndertaleResourceRef>>();
+        private Dictionary<UndertaleObject, uint> objectPool = new();
+        private Dictionary<UndertaleObject, List<uint>> pendingWrites = new();
+        private Dictionary<UndertaleObject, List<uint>> pendingStringWrites = new();
 
         public void Flush(UndertaleData data)
         {
-            SubmitMessage("Writing object references...");
-
-            var intsToWriteParallelSorted = intsToWriteParallel.AsParallel()
-                                                               .WithExecutionMode(ParallelExecutionMode.ForceParallelism)
-                                                               .OrderBy(x => x.Item1);
-            foreach (var pair in intsToWriteParallelSorted)
+            // Clear out index dictionaries (no longer needed)
+            foreach (var chunk in data.FORM.Chunks.Values)
             {
-                Position = pair.Item1;
-                Write(pair.Item2);
+                if (chunk is IUndertaleListChunk listChunk)
+                {
+                    listChunk.ClearIndexDict();
+                }
             }
 
-            var resourceIDRefsToWriteSorted = resourceIDRefsToWrite.AsParallel()
-                                                                   .WithExecutionMode(ParallelExecutionMode.ForceParallelism)
-                                                                   .OrderBy(x => x.Item1);
-            foreach (var pair in resourceIDRefsToWriteSorted)
-            {
-                int id = pair.Item2.SerializeById(this);
-                Position = pair.Item1;
-                Write(id);
-            }
-
-            SubmitMessage("Clearing temporary dictionaries...");
-            var listChunks = data.FORM.Chunks.Values.Select(x => x as IUndertaleListChunk);
-            Parallel.ForEach(listChunks.Where(x => x is not null), (chunk) =>
-            {
-                chunk.ClearIndexDict();
-            });
-            UndertaleIO.IsDictionaryCleared = true;
-
-            SubmitMessage("Flushing remaining file buffer data to disk...");
+            SubmitMessage("Flushing remaining buffer data...");
             base.Flush();
         }
 
@@ -776,31 +782,52 @@ namespace UndertaleModLib
         {
             try
             {
-                // This isn't a major issue, and this is a performance waster
-                //if (objectPool.ContainsKey(obj))
-                //    throw new IOException("Writing object twice");
+                // Store object address before writing it
                 uint objectAddr = Position;
-                if (obj.GetType() == typeof(UndertaleString))
-                {
-                    if (pendingStringWrites.ContainsKey(obj))
-                    {
-                        foreach (uint pointerAddr in pendingStringWrites[obj])
-                            intsToWriteParallel.Add(new ValueTuple<uint, uint>(pointerAddr, objectAddr + 4));
 
+                if (typeof(T) == typeof(UndertaleString))
+                {
+                    // Can skip adding strings to object pool (nothing later in the file references them).
+                    obj.Serialize(this);
+
+                    // Patch all pointers to this string, if applicable
+                    if (pendingStringWrites.TryGetValue(obj, out List<uint> patches))
+                    {
+                        uint returnTo = Position;
+                        objectAddr += 4; // Destination is where the string starts, not its length
+                        foreach (uint pointerAddr in patches)
+                        {
+                            Position = pointerAddr;
+                            Write(objectAddr);
+                        }
+                        Position = returnTo;
+
+                        // Remove pending write
                         pendingStringWrites.Remove(obj);
                     }
                 }
                 else
-                    objectPool.Add(obj, objectAddr); // strings come later in the file, so no need to add them to the pool
-                
-                obj.Serialize(this);
-
-                if (pendingWrites.ContainsKey(obj))
                 {
-                    foreach (uint pointerAddr in pendingWrites[obj])
-                        intsToWriteParallel.Add(new ValueTuple<uint, uint>(pointerAddr, objectAddr));
+                    // Add object to pool
+                    objectPool.Add(obj, objectAddr);
 
-                    pendingWrites.Remove(obj);
+                    // Serialize object
+                    obj.Serialize(this);
+
+                    // Patch all pointers to this object, if applicable
+                    if (pendingWrites.TryGetValue(obj, out List<uint> patches))
+                    {
+                        uint returnTo = Position;
+                        foreach (uint pointerAddr in patches)
+                        {
+                            Position = pointerAddr;
+                            Write(objectAddr);
+                        }
+                        Position = returnTo;
+
+                        // Remove pending write
+                        pendingWrites.Remove(obj);
+                    }
                 }
             }
             catch (Exception e)
@@ -823,9 +850,11 @@ namespace UndertaleModLib
             }
             else
             {
-                if (!pendingWrites.ContainsKey(obj))
-                    pendingWrites.Add(obj, new List<uint>());
-                pendingWrites[obj].Add(Position);
+                if (!pendingWrites.TryGetValue(obj, out List<uint> list))
+                {
+                    pendingWrites.Add(obj, list = new List<uint>());
+                }
+                list.Add(Position);
                 Write(0xDEADC0DEu);
             }
         }
@@ -838,9 +867,11 @@ namespace UndertaleModLib
                 return;
             }
 
-            if (!pendingStringWrites.ContainsKey(obj))
-                pendingStringWrites.Add(obj, new List<uint>());
-            pendingStringWrites[obj].Add(Position);
+            if (!pendingStringWrites.TryGetValue(obj, out List<uint> list))
+            {
+                pendingStringWrites.Add(obj, list = new List<uint>());
+            }
+            list.Add(Position);
             Write(0xDEADC0DEu);
         }
 
@@ -902,13 +933,11 @@ namespace UndertaleModLib
 
     public static class UndertaleIO
     {
-        public static bool IsDictionaryCleared { get; set; } = true;
-
         public static UndertaleData Read(Stream stream, UndertaleReader.WarningHandlerDelegate warningHandler = null,
                                                         UndertaleReader.MessageHandlerDelegate messageHandler = null,
                                                         bool onlyGeneralInfo = false)
         {
-            UndertaleReader reader = new UndertaleReader(stream, warningHandler, messageHandler, onlyGeneralInfo);
+            UndertaleReader reader = new(stream, warningHandler, messageHandler, onlyGeneralInfo);
             var data = reader.ReadUndertaleData();
             reader.ThrowIfUnreadObjects();
             return data;
@@ -916,7 +945,7 @@ namespace UndertaleModLib
 
         public static void Write(Stream stream, UndertaleData data, UndertaleWriter.MessageHandlerDelegate messageHandler = null)
         {
-            UndertaleWriter writer = new UndertaleWriter(stream, messageHandler);
+            UndertaleWriter writer = new(stream, messageHandler);
             writer.WriteUndertaleData(data);
             writer.ThrowIfUnwrittenObjects();
             writer.Flush(data);
@@ -924,7 +953,7 @@ namespace UndertaleModLib
 
         public static Dictionary<uint, UndertaleObject> GenerateOffsetMap(Stream stream)
         {
-            UndertaleReader reader = new UndertaleReader(stream);
+            UndertaleReader reader = new(stream);
             reader.ReadUndertaleData();
             reader.ThrowIfUnreadObjects();
             return reader.GetOffsetMap();
