@@ -29,12 +29,12 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Media.TextFormatting;
 using System.Windows.Navigation;
+using System.Windows.Shell;
 using System.Xml;
 using UndertaleModLib;
 using UndertaleModLib.Compiler;
 using UndertaleModLib.Decompiler;
 using UndertaleModLib.Models;
-using static UndertaleModTool.MainWindow.CodeEditorMode;
 using Input = System.Windows.Input;
 
 namespace UndertaleModTool
@@ -49,24 +49,21 @@ namespace UndertaleModTool
 
         public UndertaleCode CurrentDisassembled = null;
         public UndertaleCode CurrentDecompiled = null;
-        public List<string> CurrentLocals = null;
-        public string ProfileHash = mainWindow.ProfileHash;
-        public string MainPath = Path.Combine(Settings.ProfilesFolder, mainWindow.ProfileHash, "Main");
-        public string TempPath = Path.Combine(Settings.ProfilesFolder, mainWindow.ProfileHash, "Temp");
+        public List<string> CurrentLocals = new();
 
         public bool DecompiledFocused = false;
         public bool DecompiledChanged = false;
         public bool DecompiledYet = false;
         public bool DecompiledSkipped = false;
         public SearchPanel DecompiledSearchPanel;
-        public static (int Line, int Column, double ScrollPos) OverriddenDecompPos;
+        public static (int Line, int Column, double ScrollPos) OverriddenDecompPos { get; set; }
 
         public bool DisassemblyFocused = false;
         public bool DisassemblyChanged = false;
         public bool DisassembledYet = false;
         public bool DisassemblySkipped = false;
         public SearchPanel DisassemblySearchPanel;
-        public static (int Line, int Column, double ScrollPos) OverriddenDisasmPos;
+        public static (int Line, int Column, double ScrollPos) OverriddenDisasmPos { get; set; }
 
         public static RoutedUICommand Compile = new RoutedUICommand("Compile code", "Compile", typeof(UndertaleCodeEditor));
 
@@ -74,6 +71,14 @@ namespace UndertaleModTool
         private static readonly Dictionary<string, UndertaleNamedResource> ScriptsDict = new();
         private static readonly Dictionary<string, UndertaleNamedResource> FunctionsDict = new();
         private static readonly Dictionary<string, UndertaleNamedResource> CodeDict = new();
+
+        public enum CodeEditorTab
+        {
+            Unknown,
+            Disassembly,
+            Decompiled
+        }
+        public static CodeEditorTab EditorTab { get; set; } = CodeEditorTab.Unknown;
 
         public UndertaleCodeEditor()
         {
@@ -242,13 +247,16 @@ namespace UndertaleModTool
                 else
                     _ = DecompileCode(code, true);
             }
+            if (DecompiledTab.IsSelected)
+            {
+                // Re-populate local variables when in decompiled code, fixing #1320
+                PopulateCurrentLocals(mainWindow.Data, code);
+            }
         }
 
         private async void TabControl_SelectionChanged(object sender, SelectionChangedEventArgs e)
         {
             UndertaleCode code = this.DataContext as UndertaleCode;
-            Directory.CreateDirectory(MainPath);
-            Directory.CreateDirectory(TempPath);
             if (code == null)
                 return;
 
@@ -298,9 +306,9 @@ namespace UndertaleModTool
             CurrentDecompiled = null;
             CurrentDisassembled = null;
 
-            if (MainWindow.CodeEditorDecompile != Unstated) //if opened from the code search results "link"
+            if (EditorTab != CodeEditorTab.Unknown) // If opened from the code search results "link"
             {
-                if (MainWindow.CodeEditorDecompile == DontDecompile && code != CurrentDisassembled)
+                if (EditorTab == CodeEditorTab.Disassembly && code != CurrentDisassembled)
                 {
                     if (CodeModeTabs.SelectedItem != DisassemblyTab)
                         CodeModeTabs.SelectedItem = DisassemblyTab;
@@ -308,7 +316,7 @@ namespace UndertaleModTool
                         DisassembleCode(code, true);
                 }
 
-                if (MainWindow.CodeEditorDecompile == Decompile && code != CurrentDecompiled)
+                if (EditorTab == CodeEditorTab.Decompiled && code != CurrentDecompiled)
                 {
                     if (CodeModeTabs.SelectedItem != DecompiledTab)
                         CodeModeTabs.SelectedItem = DecompiledTab;
@@ -316,7 +324,7 @@ namespace UndertaleModTool
                         _ = DecompileCode(code, true);
                 }
 
-                MainWindow.CodeEditorDecompile = Unstated;
+                EditorTab = CodeEditorTab.Unknown;
             }
             else
                 FillInCodeViewer(true);
@@ -367,6 +375,9 @@ namespace UndertaleModTool
         {
             if (linePos <= textEditor.LineCount)
             {
+                if (linePos == -1)
+                    linePos = textEditor.Document.LineCount;
+
                 int lineLen = textEditor.Document.GetLineByNumber(linePos).Length;
                 textEditor.TextArea.Caret.Line = linePos;
                 if (columnPos != -1)
@@ -375,7 +386,8 @@ namespace UndertaleModTool
                     textEditor.TextArea.Caret.Column = lineLen + 1;
 
                 textEditor.ScrollToLine(linePos);
-                textEditor.ScrollToVerticalOffset(scrollPos);
+                if (scrollPos != -1)
+                    textEditor.ScrollToVerticalOffset(scrollPos);
             }
             else
             {
@@ -383,7 +395,36 @@ namespace UndertaleModTool
                 textEditor.ScrollToEnd();
             }
         }
-        
+        public static void ChangeLineNumber(int lineNum, CodeEditorTab editorTab)
+        {
+            if (lineNum < 1)
+                return;
+
+            if (editorTab == CodeEditorTab.Unknown)
+            {
+                Debug.WriteLine($"The \"{nameof(editorTab)}\" argument of \"{nameof(ChangeLineNumber)}()\" is \"{nameof(CodeEditorTab.Unknown)}\".");
+                return;
+            }
+
+            if (editorTab == CodeEditorTab.Decompiled)
+                OverriddenDecompPos = (lineNum, -1, -1);
+            else
+                OverriddenDisasmPos = (lineNum, -1, -1);
+        }
+        public static void ChangeLineNumber(int lineNum, TextEditor textEditor)
+        {
+            if (lineNum < 1)
+                return;
+
+            if (textEditor is null)
+            {
+                Debug.WriteLine($"The \"{nameof(textEditor)}\" argument of \"{nameof(ChangeLineNumber)}()\" is null.");
+                return;
+            }
+
+            RestoreCaretPosition(textEditor, lineNum, -1, -1);
+        }
+
         private static void FillObjectDicts()
         {
             var data = mainWindow.Data;
@@ -447,8 +488,6 @@ namespace UndertaleModTool
 
         private void DisassembleCode(UndertaleCode code, bool first)
         {
-            code.UpdateAddresses();
-
             string text;
 
             int currLine = 1;
@@ -474,7 +513,8 @@ namespace UndertaleModTool
             if (code.ParentEntry != null)
             {
                 DisassemblyEditor.IsReadOnly = true;
-                text = "; This code entry is a reference to an anonymous function within " + code.ParentEntry.Name.Content + ", view it there";
+                text = "; This code entry is a reference to an anonymous function within " + code.ParentEntry.Name.Content + ", view it there.";
+                DisassemblyChanged = false;
             }
             else
             {
@@ -483,9 +523,9 @@ namespace UndertaleModTool
                 try
                 {
                     var data = mainWindow.Data;
-                    text = code.Disassemble(data.Variables, data.CodeLocals.For(code));
+                    text = code.Disassemble(data.Variables, data.CodeLocals?.For(code));
 
-                    CurrentLocals = new List<string>();
+                    CurrentLocals.Clear();
                 }
                 catch (Exception ex)
                 {
@@ -513,28 +553,39 @@ namespace UndertaleModTool
         }
 
         public static Dictionary<string, string> gettext = null;
-        private void UpdateGettext(UndertaleCode gettextCode)
+        private void UpdateGettext(UndertaleData data, UndertaleCode gettextCode)
         {
             gettext = new Dictionary<string, string>();
             string[] decompilationOutput;
-            if (!SettingsWindow.ProfileModeEnabled)
-                decompilationOutput = Decompiler.Decompile(gettextCode, new GlobalDecompileContext(null, false)).Replace("\r\n", "\n").Split('\n');
+            GlobalDecompileContext context = new(data);
+            if (!SettingsWindow.ProfileModeEnabled || mainWindow.ProfileHash is not string currentMD5)
+            {
+                decompilationOutput = new Underanalyzer.Decompiler.DecompileContext(context, gettextCode, data.ToolInfo.DecompilerSettings)
+                    .DecompileToString().Split('\n');
+            }
             else
             {
-                try
+                string path = Path.Combine(MainWindow.ProfilesFolder, currentMD5, "Temp", gettextCode.Name.Content + ".gml");
+                if (File.Exists(path))
                 {
-                    string path = Path.Combine(TempPath, gettextCode.Name.Content + ".gml");
-                    if (File.Exists(path))
+                    try
+                    {
                         decompilationOutput = File.ReadAllText(path).Replace("\r\n", "\n").Split('\n');
-                    else
-                        decompilationOutput = Decompiler.Decompile(gettextCode, new GlobalDecompileContext(null, false)).Replace("\r\n", "\n").Split('\n');
+                    }
+                    catch
+                    {
+                        decompilationOutput = new Underanalyzer.Decompiler.DecompileContext(context, gettextCode, data.ToolInfo.DecompilerSettings)
+                            .DecompileToString().Split('\n');
+                    }
                 }
-                catch
+                else
                 {
-                    decompilationOutput = Decompiler.Decompile(gettextCode, new GlobalDecompileContext(null, false)).Replace("\r\n", "\n").Split('\n');
+                    decompilationOutput = new Underanalyzer.Decompiler.DecompileContext(context, gettextCode, data.ToolInfo.DecompilerSettings)
+                        .DecompileToString().Split('\n');
                 }
             }
-            Regex textdataRegex = new Regex("^ds_map_add\\(global\\.text_data_en, \\\"(.*)\\\", \\\"(.*)\\\"\\)", RegexOptions.Compiled);
+            Regex textdataRegex = new("^ds_map_add\\(global\\.text_data_en, \\\"(.*)\\\", \\\"(.*)\\\"\\)", RegexOptions.Compiled);
+            Regex textdataRegex2 = new("^ds_map_add\\(global\\.text_data_en, \\\"(.*)\\\", '(.*)'\\)", RegexOptions.Compiled);
             foreach (var line in decompilationOutput)
             {
                 Match m = textdataRegex.Match(line);
@@ -542,7 +593,14 @@ namespace UndertaleModTool
                 {
                     try
                     {
-                        gettext.Add(m.Groups[1].Value, m.Groups[2].Value);
+                        if (!data.IsGameMaker2() && m.Groups[2].Value.Contains("'\"'"))
+                        {
+                            gettext.Add(m.Groups[1].Value, $"\"{m.Groups[2].Value}\"");
+                        }
+                        else
+                        {
+                            gettext.Add(m.Groups[1].Value, m.Groups[2].Value);
+                        }
                     }
                     catch (ArgumentException)
                     {
@@ -553,12 +611,38 @@ namespace UndertaleModTool
                         mainWindow.ShowError("Unknown error in textdata_en. This may cause errors in the comment display of text.");
                     }
                 }
+                else
+                {
+                    m = textdataRegex2.Match(line);
+                    if (m.Success)
+                    {
+                        try
+                        {
+                            if (!data.IsGameMaker2() && m.Groups[2].Value.Contains("\"'\""))
+                            {
+                                gettext.Add(m.Groups[1].Value, $"\'{m.Groups[2].Value}\'");
+                            }
+                            else
+                            {
+                                gettext.Add(m.Groups[1].Value, m.Groups[2].Value);
+                            }
+                        }
+                        catch (ArgumentException)
+                        {
+                            mainWindow.ShowError("There is a duplicate key in textdata_en, being " + m.Groups[1].Value + ". This may cause errors in the comment display of text.");
+                        }
+                        catch
+                        {
+                            mainWindow.ShowError("Unknown error in textdata_en. This may cause errors in the comment display of text.");
+                        }
+                    }
+                }
             }
         }
 
         public static Dictionary<string, string> gettextJSON = null;
-        private static readonly Regex gettextRegex = new("scr_gettext\\(\\\"(.*?)\\\"\\)", RegexOptions.Compiled);
-        private static readonly Regex getlangRegex = new("scr_84_get_lang_string(?:.*?)\\(\\\"(.*?)\\\"\\)", RegexOptions.Compiled);
+        private static readonly Regex gettextRegex = new(@"scr_gettext\(\""(.*?)\""\)(?!(.*?\/\/.*?$))", RegexOptions.Compiled);
+        private static readonly Regex getlangRegex = new(@"scr_84_get_lang_string(?:.*?)\(\""(.*?)\""\)(?!(.*?\/\/.*?$))", RegexOptions.Compiled);
         private string UpdateGettextJSON(string json)
         {
             try
@@ -600,7 +684,8 @@ namespace UndertaleModTool
 
             if (code.ParentEntry != null)
             {
-                DecompiledEditor.Text = "// This code entry is a reference to an anonymous function within " + code.ParentEntry.Name.Content + ", view it there";
+                DecompiledEditor.Text = "// This code entry is a reference to an anonymous function within " + code.ParentEntry.Name.Content + ", view it there.";
+                DecompiledChanged = false;
                 CurrentDecompiled = code;
                 existingDialog?.TryClose();
             }
@@ -644,18 +729,29 @@ namespace UndertaleModTool
                 var dataa = mainWindow.Data;
                 Task t = Task.Run(() =>
                 {
-                    GlobalDecompileContext context = new GlobalDecompileContext(dataa, false);
+                    GlobalDecompileContext context = new(dataa);
                     string decompiled = null;
                     Exception e = null;
                     try
                     {
-                        string path = Path.Combine(TempPath, code.Name.Content + ".gml");
-                        if (!SettingsWindow.ProfileModeEnabled || !File.Exists(path))
+                        if (!SettingsWindow.ProfileModeEnabled || mainWindow.ProfileHash is not string currentMD5)
                         {
-                            decompiled = Decompiler.Decompile(code, context, (msg) => { dialog.Message = msg; });
+                            decompiled = new Underanalyzer.Decompiler.DecompileContext(context, code, dataa.ToolInfo.DecompilerSettings)
+                                .DecompileToString();
                         }
                         else
-                            decompiled = File.ReadAllText(path);
+                        {
+                            string path = Path.Combine(MainWindow.ProfilesFolder, currentMD5, "Temp", code.Name.Content + ".gml");
+                            if (!File.Exists(path))
+                            {
+                                decompiled = new Underanalyzer.Decompiler.DecompileContext(context, code, dataa.ToolInfo.DecompilerSettings)
+                                    .DecompileToString();
+                            }
+                            else
+                            {
+                                decompiled = File.ReadAllText(path);
+                            }
+                        }
                     }
                     catch (Exception ex)
                     {
@@ -663,7 +759,7 @@ namespace UndertaleModTool
                     }
 
                     if (gettextCode != null)
-                        UpdateGettext(gettextCode);
+                        UpdateGettext(dataa, gettextCode);
 
                     try
                     {
@@ -715,7 +811,7 @@ namespace UndertaleModTool
                                     {
                                         Match match = matches[i];
                                         if (!currDict.TryGetValue(match.Groups[1].Value, out string text))
-                                            continue;
+                                            text = "<localization fetch error>";
 
                                         if (i != matches.Length - 1) // If not the last
                                             decompLinesBuilder.Append($"{text}; ");
@@ -744,22 +840,12 @@ namespace UndertaleModTool
                         else if (decompiled != null)
                         {
                             DecompiledEditor.Document.Text = decompiled;
-                            CurrentLocals = new List<string>();
-
-                            var locals = dataa.CodeLocals.ByName(code.Name.Content);
-                            if (locals != null)
-                            {
-                                foreach (var local in locals.Locals)
-                                    CurrentLocals.Add(local.Name.Content);
-                            }
+                            PopulateCurrentLocals(dataa, code);
 
                             RestoreCaretPosition(DecompiledEditor, currLine, currColumn, scrollPos);
 
                             if (existingDialog is not null)                      //if code was edited (and compiles after it)
                             {
-                                dataa.GMLCacheChanged.Add(code.Name.Content);
-                                dataa.GMLCacheFailed?.Remove(code.Name.Content); //remove that code name, since that code compiles now
-
                                 openSaveDialog = mainWindow.IsSaving;
                             }
                         }
@@ -782,6 +868,19 @@ namespace UndertaleModTool
 
                 if (openSaveDialog)
                     await mainWindow.DoSaveDialog();
+            }
+        }
+
+        private void PopulateCurrentLocals(UndertaleData data, UndertaleCode code)
+        {
+            CurrentLocals.Clear();
+
+            // Look up locals for given code entry's name, for syntax highlighting
+            var locals = data.CodeLocals?.ByName(code.Name.Content);
+            if (locals != null)
+            {
+                foreach (var local in locals.Locals)
+                    CurrentLocals.Add(local.Name.Content);
             }
         }
 
@@ -849,68 +948,52 @@ namespace UndertaleModTool
                 // This is still a problem in rare cases for some unknown reason
             }
 
-            CompileContext compileContext = null;
+            CompileResult compileResult = new();
+            string rootException = null;
             string text = DecompiledEditor.Text;
             var dispatcher = Dispatcher;
             Task t = Task.Run(() =>
             {
                 try
                 {
-                    compileContext = Compiler.CompileGMLText(text, data, code, (f) => { dispatcher.Invoke(() => f()); });
+                    CompileGroup group = new(data);
+                    group.MainThreadAction = (f) => { dispatcher.Invoke(() => f()); };
+                    group.QueueCodeReplace(code, text);
+                    compileResult = group.Compile();
                 }
                 catch (Exception ex)
                 {
-                    compileContext = new(data, code)
-                    {
-                        HasError = true,
-                        ResultError = ex.ToString()
-                    };
+                    rootException = ex.ToString();
                 }
             });
             await t;
 
-            if (compileContext == null)
+            if (rootException is not null)
             {
                 dialog.TryClose();
-                mainWindow.ShowError("Compile context was null for some reason...", "This shouldn't happen");
+                mainWindow.ShowError(Truncate(rootException, 512), "Compiler error");
                 return;
             }
 
-            if (compileContext.HasError)
+            if (!compileResult.Successful)
             {
                 dialog.TryClose();
-                mainWindow.ShowError(Truncate(compileContext.ResultError, 512), "Compiler error");
+                mainWindow.ShowError(Truncate(compileResult.PrintAllErrors(false), 512), "Compiler error");
                 return;
             }
 
-            if (!compileContext.SuccessfulCompile)
+            if (SettingsWindow.ProfileModeEnabled && mainWindow.ProfileHash is string currentMD5)
             {
-                dialog.TryClose();
-                mainWindow.ShowError("(unknown error message)", "Compile failed");
-                return;
-            }
-
-            code.Replace(compileContext.ResultAssembly);
-            try
-            {
-                string path = Path.Combine(TempPath, code.Name.Content + ".gml");
-                if (SettingsWindow.ProfileModeEnabled)
+                try
                 {
                     // Write text, only if in the profile mode.
+                    string path = Path.Combine(MainWindow.ProfilesFolder, currentMD5, "Temp", code.Name.Content + ".gml");
                     File.WriteAllText(path, DecompiledEditor.Text);
                 }
-                else
+                catch (Exception exc)
                 {
-                    // Destroy file with comments if it's been edited outside the profile mode.
-                    // We're dealing with the decompiled code only, it has to happen.
-                    // Otherwise it will cause a desync, which is more important to prevent.
-                    if (File.Exists(path))
-                        File.Delete(path);
+                    mainWindow.ShowError("Error during writing of GML code to profile:\n" + exc);
                 }
-            }
-            catch (Exception exc)
-            {
-                mainWindow.ShowError("Error during writing of GML code to profile:\n" + exc);
             }
 
             // Invalidate gettext if necessary
@@ -930,8 +1013,6 @@ namespace UndertaleModTool
 
             // Decompile new code
             await DecompileCode(code, false, dialog);
-
-            //GMLCacheChanged.Add() is inside DecompileCode()
         }
         private void DecompiledEditor_LostFocus(object sender, RoutedEventArgs e)
         {
@@ -986,7 +1067,6 @@ namespace UndertaleModTool
             {
                 var instructions = Assembler.Assemble(DisassemblyEditor.Text, data);
                 code.Replace(instructions);
-                mainWindow.NukeProfileGML(code.Name.Content);
             }
             catch (Exception ex)
             {
@@ -1007,8 +1087,6 @@ namespace UndertaleModTool
 
             if (!DisassemblyEditor.IsReadOnly)
             {
-                data.GMLCacheChanged.Add(code.Name.Content);
-
                 if (mainWindow.IsSaving)
                 {
                     mainWindow.IsSaving = false;
@@ -1300,6 +1378,7 @@ namespace UndertaleModTool
                 bool func = (offset + nameLength + 1 < CurrentContext.VisualLine.LastDocumentLine.EndOffset) &&
                             (doc.GetCharAt(offset + nameLength) == '(');
                 UndertaleNamedResource val = null;
+                bool nonResourceReference = false;
 
                 var editor = textEditorInst;
 
@@ -1322,7 +1401,7 @@ namespace UndertaleModTool
                             else
                             {
                                 // Resolve 2.3 sub-functions for their parent entry
-                                if (data.KnownSubFunctions?.TryGetValue(nameText, out UndertaleFunction f) == true)
+                                if (data.GlobalFunctions?.TryGetFunction(nameText, out Underanalyzer.IGMFunction f) == true)
                                 {
                                     ScriptsDict.TryGetValue(f.Name.Content, out val);
                                     val = (val as UndertaleScript)?.Code?.ParentEntry;
@@ -1344,11 +1423,50 @@ namespace UndertaleModTool
                 else
                 {
                     NamedObjDict.TryGetValue(nameText, out val);
-                    if (data.IsVersionAtLeast(2, 3) & val is UndertaleScript)
-                        val = null; // in GMS2.3 scripts are never referenced directly
+                    if (data.IsVersionAtLeast(2, 3))
+                    { 
+                        if (val is UndertaleScript)
+                            val = null; // in GMS2.3 scripts are never referenced directly
+
+                        if (data.GlobalFunctions?.TryGetFunction(nameText, out Underanalyzer.IGMFunction globalFunc) == true &&
+                            globalFunc is UndertaleFunction utGlobalFunc)
+                        {
+                            // Try getting script that this function reference belongs to
+                            if (NamedObjDict.TryGetValue("gml_Script_" + nameText, out val) && val is UndertaleScript script)
+                            {
+                                // Highlight like a function as well
+                                val = script.Code;
+                                func = true;
+                            }
+                        }
+
+                        if (val == null)
+                        {
+                            // Try to get basic function
+                            if (FunctionsDict.TryGetValue(nameText, out val))
+                            {
+                                func = true;
+                            }
+                        }
+
+                        if (val == null)
+                        {
+                            // Try resolving to room instance ID
+                            string instanceIdPrefix = data.ToolInfo.InstanceIdPrefix();
+                            if (nameText.StartsWith(instanceIdPrefix) &&
+                                int.TryParse(nameText[instanceIdPrefix.Length..], out int id) && id >= 100000)
+                            {
+                                // TODO: We currently mark this as a non-resource reference, but ideally
+                                // we resolve this to the room that this instance ID occurs in.
+                                // However, we should only do this when actually clicking on it.
+                                nonResourceReference = true;
+                            }
+                        }
+                    }
                 }
-                if (val == null)
+                if (val == null && !nonResourceReference)
                 {
+                    // Check for variable name colors
                     if (offset >= 7)
                     {
                         if (doc.GetText(offset - 7, 7) == "global.")
@@ -1374,19 +1492,26 @@ namespace UndertaleModTool
                 var line = new ClickVisualLineText(nameText, CurrentContext.VisualLine, nameLength,
                                                    func ? FunctionBrush : ConstantBrush);
                 if (func)
-                    line.Bold = true;
-                line.Clicked += async (text, button) =>
                 {
-                    await codeEditorInst?.SaveChanges();
-
-                    if (button == Input.MouseButton.Right)
+                    // Make function references bold as well as a different color
+                    line.Bold = true;
+                }
+                if (val is not null)
+                {
+                    // Add click operation when we have a resource
+                    line.Clicked += async (text, button) =>
                     {
-                        contextMenu.DataContext = val;
-                        contextMenu.IsOpen = true;
-                    }
-                    else
-                        mainWindow.ChangeSelection(val, button == Input.MouseButton.Middle);
-                };
+                        await codeEditorInst?.SaveChanges();
+
+                        if (button == Input.MouseButton.Right)
+                        {
+                            contextMenu.DataContext = val;
+                            contextMenu.IsOpen = true;
+                        }
+                        else
+                            mainWindow.ChangeSelection(val, button == Input.MouseButton.Middle);
+                    };
+                }
 
                 return line;
             }
