@@ -2,7 +2,6 @@ using System.IO;
 
 namespace UndertaleModCli;
 
-// ReSharper disable NotNullMemberIsNotInitialized - Properties are applied via reflection.
 /// <summary>
 /// Cli options for the Load command
 /// </summary>
@@ -21,12 +20,12 @@ public class LoadOptions
 	/// <summary>
 	/// C# string that shall be executed
 	/// </summary>
-	public string? Line { get; set; }
+	public string Line { get; set; }
 
 	/// <summary>
 	/// File path to where to save the modified data file
 	/// </summary>
-	public FileInfo? Output { get; set; }
+	public FileInfo Output { get; set; }
 
 	/// <summary>
 	/// Determines if Cli should be run in interactive mode
