@@ -2,7 +2,6 @@ using System.IO;
 
 namespace UndertaleModCli;
 
-// ReSharper disable NotNullMemberIsNotInitialized - Properties are applied via reflection.
 /// <summary>
 /// Cli options for the Replace command
 /// </summary>
@@ -16,7 +15,7 @@ public class ReplaceOptions
     /// <summary>
     /// File path to where to save the modified data file
     /// </summary>
-    public FileInfo? Output { get; set; }
+    public FileInfo Output { get; set; }
 
     /// <summary>
     /// Determines if Cli should print out verbose logs
