@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
+using Underanalyzer;
 using static UndertaleModLib.Models.UndertaleGeneralInfo;
 
 namespace UndertaleModLib.Models;
@@ -10,7 +11,7 @@ namespace UndertaleModLib.Models;
 /// A function entry as it's used in a GameMaker data file.
 /// </summary>
 [PropertyChanged.AddINotifyPropertyChangedInterface]
-public class UndertaleFunction : UndertaleNamedResource, UndertaleInstruction.ReferencedObject, IStaticChildObjectsSize, IDisposable
+public class UndertaleFunction : UndertaleNamedResource, UndertaleInstruction.IReferencedObject, IStaticChildObjectsSize, IDisposable, IGMFunction
 {
     /// <inheritdoc cref="IStaticChildObjectsSize.ChildObjectsSize" />
     public static readonly uint ChildObjectsSize = 12;
@@ -31,9 +32,11 @@ public class UndertaleFunction : UndertaleNamedResource, UndertaleInstruction.Re
     /// How often this <see cref="UndertaleFunction"/> is referenced in code.
     /// </summary>
     public uint Occurrences { get; set; }
+    
+    /// <summary>
+    /// The first instruction in this function.
+    /// </summary>
     public UndertaleInstruction FirstAddress { get; set; }
-
-    public bool Autogenerated { get; set; } = false; // Whether a gml_Script_... kind of function in 2.3
 
     [Obsolete("This variable has been renamed to NameStringID.")]
     public int UnknownChainEndingValue { get => NameStringID; set => NameStringID = value; }
@@ -66,7 +69,7 @@ public class UndertaleFunction : UndertaleNamedResource, UndertaleInstruction.Re
                 FirstAddress = reader.GetUndertaleObjectAtAddress<UndertaleInstruction>(reader.ReadUInt32() - 4);
             else
                 FirstAddress = reader.ReadUndertaleObjectPointer<UndertaleInstruction>();
-            UndertaleInstruction.Reference<UndertaleFunction>.ParseReferenceChain(reader, this);
+            UndertaleInstruction.ParseReferenceChain(reader, this);
         }
         else
         {
@@ -90,15 +93,46 @@ public class UndertaleFunction : UndertaleNamedResource, UndertaleInstruction.Re
         Name = null;
         FirstAddress = null;
     }
+
+    // Underanalyzer implementations
+    IGMString IGMFunction.Name => Name;
 }
 
-// Seems to be unused. You can remove all entries and the game still works normally.
+// Seems to be unused. You can remove all entries and the game still works normally. TODO: not true, incorrect/missing locals can wreak some weird havoc like random segfaults.
 // Maybe the GM:S debugger uses this data?
 // TODO: INotifyPropertyChanged
 public class UndertaleCodeLocals : UndertaleNamedResource, IDisposable
 {
+    /// <summary>
+    /// The name of the code local entry which corresponds to a code entry.
+    /// </summary>
     public UndertaleString Name { get; set; }
+    
+    /// <summary>
+    /// A collection of local variables.
+    /// </summary>
     public ObservableCollection<LocalVar> Locals { get; private set; } = new ObservableCollection<LocalVar>();
+
+    /// <summary>
+    /// Creates an empty code locals entry with the given name.
+    /// </summary>
+    /// <param name="data">Data to add the new code to.</param>
+    /// <param name="name">Name of the new code locals entry to create.</param>
+    /// <returns>The new code locals entry.</returns>
+    public static UndertaleCodeLocals CreateEmptyEntry(UndertaleData data, UndertaleString name)
+    {
+        UndertaleCodeLocals locals = new()
+        {
+            Name = name
+        };
+        locals.Locals.Add(new()
+        {
+            Name = data.Strings.MakeString("arguments", out int argumentsStringId),
+            Index = data.IsVersionAtLeast(2, 3) ? (uint)argumentsStringId : 0
+        });
+        data.CodeLocals.Add(locals);
+        return locals;
+    }
 
     /// <inheritdoc />
     public void Serialize(UndertaleWriter writer)
@@ -133,18 +167,33 @@ public class UndertaleCodeLocals : UndertaleNamedResource, IDisposable
         return count;
     }
 
+    /// <summary>
+    /// Returns whether this code local entry contains any local variables matching a name.
+    /// </summary>
+    /// <param name="varName">The variable name to search for.</param>
+    /// <returns><see langword="true"/> if a variable matching the provided name exists, otherwise <see langword="false"/>.</returns>
     public bool HasLocal(string varName)
     {
         return Locals.Any(local=>local.Name.Content == varName);
     }
 
     // TODO: INotifyPropertyChanged
+    /// <summary>
+    /// A local variable. TODO: a better description for this.
+    /// </summary>
     public class LocalVar : UndertaleObject, IStaticChildObjectsSize, IDisposable
     {
         /// <inheritdoc cref="IStaticChildObjectsSize.ChildObjectsSize" />
         public static readonly uint ChildObjectsSize = 8;
 
+        /// <summary>
+        /// TODO: have no idea what this index does.
+        /// </summary>
         public uint Index { get; set; }
+        
+        /// <summary>
+        /// The name of the local variable.
+        /// </summary>
         public UndertaleString Name { get; set; }
 
         /// <inheritdoc />
