@@ -206,6 +206,21 @@ public class UndertaleGeneralInfo : UndertaleObject, IDisposable
     public UndertaleString Name { get; set; }
 
     /// <summary>
+    /// Different GameMaker release branches. LTS has some but not all features of equivalent newer versions.
+    /// </summary>
+    public enum BranchType
+    {
+        Pre2022_0,
+        LTS2022_0,
+        Post2022_0
+    }
+
+    /// <summary>
+    /// The GameMaker release branch of the data file. May be set to <see cref="BranchType.Post2022_0"/> when features exempted from LTS are detected.
+    /// </summary>
+    public BranchType Branch = BranchType.Pre2022_0;
+
+    /// <summary>
     /// The major version of the data file.
     /// If greater than 1, serialization produces "2.0.0.0" due to the flag no longer updating in data.win
     /// </summary>
@@ -309,33 +324,22 @@ public class UndertaleGeneralInfo : UndertaleObject, IDisposable
     /// </summary>
     public bool InfoTimestampOffset { get; set; } = true;
 
-    public static (uint, uint, uint, uint) TestForCommonGMSVersions(UndertaleReader reader,
-                                                                    (uint, uint, uint, uint) readVersion)
+    public static (uint, uint, uint, uint, BranchType) TestForCommonGMSVersions(UndertaleReader reader,
+                                                                    (uint, uint, uint, uint, BranchType) readVersion)
     {
-        (uint Major, uint Minor, uint Release, uint Build) detectedVer = readVersion;
+        (uint Major, uint Minor, uint Release, uint Build, BranchType Branch) detectedVer = readVersion;
 
         // Some GMS2+ version detection. The rest is spread around, mostly in UndertaleChunks.cs
-        if (reader.AllChunkNames.Contains("PSEM"))      // 2023.2
-            detectedVer = (2023, 2, 0, 0);
+        if (reader.AllChunkNames.Contains("PSEM"))      // 2023.2, not present on LTS
+            detectedVer = (2023, 2, 0, 0, BranchType.Post2022_0);
         else if (reader.AllChunkNames.Contains("FEAT")) // 2022.8
-            detectedVer = (2022, 8, 0, 0);
+            detectedVer = (2022, 8, 0, 0, BranchType.Pre2022_0);
         else if (reader.AllChunkNames.Contains("FEDS")) // 2.3.6
-            detectedVer = (2, 3, 6, 0);
+            detectedVer = (2, 3, 6, 0, BranchType.Pre2022_0);
         else if (reader.AllChunkNames.Contains("SEQN")) // 2.3
-            detectedVer = (2, 3, 0, 0);
+            detectedVer = (2, 3, 0, 0, BranchType.Pre2022_0);
         else if (reader.AllChunkNames.Contains("TGIN")) // 2.2.1
-            detectedVer = (2, 2, 1, 0);
-
-        if (detectedVer.Major > 2 || (detectedVer.Major == 2 && detectedVer.Minor >= 3))
-        {
-            CompileContext.GMS2_3 = true;
-            DecompileContext.GMS2_3 = true;
-        }
-        else
-        {
-            CompileContext.GMS2_3 = false;
-            DecompileContext.GMS2_3 = false;
-        }
+            detectedVer = (2, 2, 1, 0, BranchType.Pre2022_0);
 
         return detectedVer;
     }
@@ -466,24 +470,13 @@ public class UndertaleGeneralInfo : UndertaleObject, IDisposable
         if (reader.ReadOnlyGEN8)
             return;
 
-        var detectedVer = TestForCommonGMSVersions(reader, (Major, Minor, Release, Build));
-        (Major, Minor, Release, Build) = detectedVer;
-
-        if (reader.undertaleData.GeneralInfo is not null)
-        {
-            var prevGenInfo = reader.undertaleData.GeneralInfo;
-            // If previous version is greater than current
-            if (prevGenInfo.Major > Major
-                || prevGenInfo.Major == Major && prevGenInfo.Minor > Minor
-                || prevGenInfo.Major == Major && prevGenInfo.Minor == Minor && prevGenInfo.Release > Release
-                || prevGenInfo.Major == Major && prevGenInfo.Minor == Minor && prevGenInfo.Release == Release && prevGenInfo.Build > Build)
-            {
-                Major = prevGenInfo.Major;
-                Minor = prevGenInfo.Minor;
-                Release = prevGenInfo.Release;
-                Build = prevGenInfo.Build;
-            }
-        }
+        // TestForCommonGMSVersions is run during the object counting phase, so the previous general info is always accurate.
+        var prevGenInfo = reader.undertaleData.GeneralInfo;
+        Major = prevGenInfo.Major;
+        Minor = prevGenInfo.Minor;
+        Release = prevGenInfo.Release;
+        Build = prevGenInfo.Build;
+        Branch = prevGenInfo.Branch;
 
         DefaultWindowWidth = reader.ReadUInt32();
         DefaultWindowHeight = reader.ReadUInt32();
@@ -590,9 +583,16 @@ public class UndertaleGeneralInfo : UndertaleObject, IDisposable
                 sb.Append(" (GMS ");
             else
                 sb.Append(" (GM ");
-            sb.Append(Major);
-            sb.Append('.');
-            sb.Append(Minor);
+            if (Branch == BranchType.LTS2022_0) // TODO: Is there some way to dynamically get this from the enum?
+            {
+                sb.Append("2022.0");
+            }
+            else
+            {
+                sb.Append(Major);
+                sb.Append('.');
+                sb.Append(Minor);
+            }
             if (Release != 0)
             {
                 sb.Append('.');
@@ -692,7 +692,8 @@ public class UndertaleOptions : UndertaleObject, IDisposable
         UseRearTouch = 0x2000000,
         UseFastCollision = 0x4000000,
         FastCollisionCompatibility = 0x8000000,
-        DisableSandbox = 0x10000000
+        DisableSandbox = 0x10000000,
+        EnableCopyOnWrite = 0x20000000
     }
 
     /// <summary>
