@@ -6,7 +6,8 @@ namespace UndertaleModLib.Models;
 /// <summary>
 /// A global initialization entry in a data file.
 /// </summary>
-/// <remarks>Never seen in GMS1.4 so uncertain if the structure was the same.</remarks>
+/// <remarks></remarks>
+// TODO: Never seen in GMS1.4 so uncertain if the structure was the same.
 public class UndertaleGlobalInit : UndertaleObject, INotifyPropertyChanged, IDisposable
 {
     private UndertaleResourceById<UndertaleCode, UndertaleChunkCODE> _code = new();
@@ -14,6 +15,7 @@ public class UndertaleGlobalInit : UndertaleObject, INotifyPropertyChanged, IDis
     /// <summary>
     /// The <see cref="UndertaleCode"/> object which contains the code.
     /// </summary>
+    /// <remarks>This code is executed at a global scope, before the first room of the game executes.</remarks>
     public UndertaleCode Code { get => _code.Resource; set { _code.Resource = value; PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(Code))); } }
 
     /// <inheritdoc />
@@ -29,7 +31,7 @@ public class UndertaleGlobalInit : UndertaleObject, INotifyPropertyChanged, IDis
     public void Unserialize(UndertaleReader reader)
     {
         _code = new UndertaleResourceById<UndertaleCode, UndertaleChunkCODE>();
-        _code.Unserialize(reader); // TODO: reader.ReadUndertaleObject if one object starts with another one
+        _code.Unserialize(reader);   // Cannot use ReadUndertaleObject, as that messes things up.
     }
 
     /// <inheritdoc/>
