@@ -28,10 +28,15 @@ public class UndertaleSpineTextureEntry : UndertaleObject, IDisposable
     public int PageHeight { get; set; }
     
     /// <summary>
-    /// The atlas as raw bytes, can be a GameMaker QOI texture or a PNG file.
+    /// The atlas as raw bytes, can be a GameMaker QOI texture or a PNG file. Null for versions >= 2023.1.
     /// </summary>
     public byte[] TexBlob { get; set; }
-    
+
+    /// <summary>
+    /// The length of the corresponding sprite texture entry for versions >= 2023.1.
+    /// </summary>
+    public int TextureEntryLength { get; set; }
+
     /// <summary>
     /// Indicates whether <see cref="TexBlob"/> contains a GameMaker QOI texture (the header is qoif reversed).
     /// </summary>
@@ -42,8 +47,15 @@ public class UndertaleSpineTextureEntry : UndertaleObject, IDisposable
     {
         writer.Write(PageWidth);
         writer.Write(PageHeight);
-        writer.Write(TexBlob.Length);
-        writer.Write(TexBlob);
+        if (writer.undertaleData.IsVersionAtLeast(2023, 1))
+        {
+            writer.Write(TextureEntryLength);
+        } 
+        else 
+        {
+            writer.Write(TexBlob.Length);
+            writer.Write(TexBlob);
+        }
     }
 
     /// <inheritdoc />
@@ -51,14 +63,20 @@ public class UndertaleSpineTextureEntry : UndertaleObject, IDisposable
     {
         PageWidth = reader.ReadInt32();
         PageHeight = reader.ReadInt32();
-        TexBlob = reader.ReadBytes(reader.ReadInt32());
+        if (reader.undertaleData.IsVersionAtLeast(2023, 1))
+            TextureEntryLength = reader.ReadInt32();
+        else
+            TexBlob = reader.ReadBytes(reader.ReadInt32());
     }
 
     /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
     public static uint UnserializeChildObjectCount(UndertaleReader reader)
     {
         reader.Position += 8;                        // Size
-        reader.Position += (uint)reader.ReadInt32(); // "TexBlob"
+        if (reader.undertaleData.IsVersionAtLeast(2023, 1))
+            reader.Position += 4; // "TextureEntryLength"
+        else
+            reader.Position += (uint)reader.ReadInt32(); // "TexBlob"
 
         return 0;
     }
@@ -214,6 +232,7 @@ public class UndertaleSprite : UndertaleNamedResource, PrePaddedObject, INotifyP
     public UndertaleSimpleList<UndertaleSpineTextureEntry> SpineTextures { get; set; }
 
     public bool IsSpineSprite { get => SpineJSON != null && SpineAtlas != null && SpineTextures != null; }
+    public bool SpineHasTextureData { get; set; } = true;
     public bool IsYYSWFSprite { get => YYSWF != null; }
 
     private int _SWFVersion;
@@ -258,12 +277,27 @@ public class UndertaleSprite : UndertaleNamedResource, PrePaddedObject, INotifyP
         V3NineSlice = null;
     }
 
-    public MaskEntry NewMaskEntry()
+    /// <summary>
+    /// Creates a new mask entry for this sprite.
+    /// </summary>
+    /// <param name="data">Data that this sprite is part of, for checking the GameMaker version.</param>
+    /// <returns>The new mask entry.</returns>
+    public MaskEntry NewMaskEntry(UndertaleData data = null)
     {
-        MaskEntry newEntry = new MaskEntry();
-        uint len = (Width + 7) / 8 * Height;
-        newEntry.Data = new byte[len];
-        return newEntry;
+        int width, height;
+        if (data is not null)
+        {
+            // Support for 2024.6+ (modern code path)
+            (width, height) = CalculateMaskDimensions(data);
+        }
+        else
+        {
+            // Legacy code path (for scripts that haven't been updated to support 2024.6+ yet)
+            (width, height) = ((int)Width, (int)Height);
+        }
+
+        uint len = (uint)((width + 7) / 8 * height);
+        return new MaskEntry(new byte[len], width, height);
     }
 
     /// <summary>
@@ -329,13 +363,24 @@ public class UndertaleSprite : UndertaleNamedResource, PrePaddedObject, INotifyP
     {
         public byte[] Data { get; set; }
 
+        /// <summary>
+        /// Width of this sprite mask. UTMT only.
+        /// </summary>
+        public int Width { get; set; }
+        /// <summary>
+        /// Height of this sprite mask. UTMT only.
+        /// </summary>
+        public int Height { get; set; }
+
         public MaskEntry()
         {
         }
 
-        public MaskEntry(byte[] data)
+        public MaskEntry(byte[] data, int width, int height)
         {
             this.Data = data;
+            this.Width = width;
+            this.Height = height;
         }
 
         /// <inheritdoc/>
@@ -404,6 +449,9 @@ public class UndertaleSprite : UndertaleNamedResource, PrePaddedObject, INotifyP
 
                     byte[] encodedJson = EncodeSpineBlob(Encoding.UTF8.GetBytes(SpineJSON));
                     byte[] encodedAtlas = EncodeSpineBlob(Encoding.UTF8.GetBytes(SpineAtlas));
+
+                    if (writer.undertaleData.IsVersionAtLeast(2023, 1))
+                        writer.WriteUndertaleObject(Textures);
 
                     // the header.
                     writer.Write(SpineVersion);
@@ -492,7 +540,9 @@ public class UndertaleSprite : UndertaleNamedResource, PrePaddedObject, INotifyP
             writer.Write((byte)0);
             total++;
         }
-        Util.DebugUtil.Assert(total == CalculateMaskDataSize(Width, Height, (uint)CollisionMasks.Count), "Invalid mask data for sprite");
+
+        (int width, int height) = CalculateMaskDimensions(writer.undertaleData);
+        Util.DebugUtil.Assert(total == CalculateMaskDataSize(width, height, (uint)CollisionMasks.Count), "Invalid mask data for sprite");
     }
 
     private static byte[] DecodeSpineBlob(byte[] blob)
@@ -585,6 +635,12 @@ public class UndertaleSprite : UndertaleNamedResource, PrePaddedObject, INotifyP
                 {
                     reader.Align(4);
 
+                    if (reader.undertaleData.IsVersionAtLeast(2023, 1))
+                    {
+                        Textures = reader.ReadUndertaleObject<UndertaleSimpleList<TextureEntry>>();
+                        SpineHasTextureData = false;
+                    }
+
                     SpineVersion = reader.ReadInt32();
                     if (SpineVersion >= 3)
                     {
@@ -669,8 +725,12 @@ public class UndertaleSprite : UndertaleNamedResource, PrePaddedObject, INotifyP
         reader.Position += 4; // "Name"
         uint width = reader.ReadUInt32();
         uint height = reader.ReadUInt32();
+        int marginLeft = reader.ReadInt32();
+        int marginRight = reader.ReadInt32();
+        int marginBottom = reader.ReadInt32();
+        int marginTop = reader.ReadInt32();
 
-        reader.Position += 44;
+        reader.Position += 28;
 
         if (reader.ReadInt32() == -1)
         {
@@ -699,7 +759,7 @@ public class UndertaleSprite : UndertaleNamedResource, PrePaddedObject, INotifyP
             {
                 case SpriteType.Normal:
                     count += 1 + UndertaleSimpleList<TextureEntry>.UnserializeChildObjectCount(reader);
-                    SkipMaskData(reader, width, height);
+                    SkipMaskData(reader, (int)width, (int)height, marginRight, marginLeft, marginBottom, marginTop);
                     break;
 
                 case SpriteType.SWF:
@@ -714,6 +774,9 @@ public class UndertaleSprite : UndertaleNamedResource, PrePaddedObject, INotifyP
                 {
                     reader.Align(4);
 
+                    if (reader.undertaleData.IsVersionAtLeast(2023, 1))
+                        count += 1 + UndertaleSimpleList<TextureEntry>.UnserializeChildObjectCount(reader);
+
                     int spineVersion = reader.ReadInt32();
                     if (spineVersion >= 3)
                         reader.Position += 4; // "SpineCacheVersion"
@@ -727,13 +790,13 @@ public class UndertaleSprite : UndertaleNamedResource, PrePaddedObject, INotifyP
                     switch (spineVersion)
                     {
                         case 1:
-                            reader.Position += 8 + jsonLength + atlasLength + textures;
+                            reader.Position += 8 + (uint)jsonLength + (uint)atlasLength + (uint)textures;
                             break;
 
                         case 2:
                         case 3:
                         {
-                            reader.Position += jsonLength + atlasLength;
+                            reader.Position += (uint)jsonLength + (uint)atlasLength;
 
                             // TODO: make this return count instead if spine sprite
                             // couldn't have sequence or nine slices data.
@@ -765,38 +828,90 @@ public class UndertaleSprite : UndertaleNamedResource, PrePaddedObject, INotifyP
         {
             reader.Position -= 4;
             count += 1 + UndertaleSimpleList<TextureEntry>.UnserializeChildObjectCount(reader);
-            SkipMaskData(reader, width, height);
+            SkipMaskData(reader, (int)width, (int)height, marginRight, marginLeft, marginBottom, marginTop);
         }
 
         return count;
     }
 
+    /// <summary>
+    /// Returns the width and height of the collision mask for this sprite, which changes depending on GameMaker version.
+    /// </summary>
+    public (int Width, int Height) CalculateMaskDimensions(UndertaleData data)
+    {
+        if (data.IsVersionAtLeast(2024, 6))
+        {
+            return CalculateBboxMaskDimensions(MarginRight, MarginLeft, MarginBottom, MarginTop);
+        }
+        return CalculateFullMaskDimensions((int)Width, (int)Height);
+    }
+
+    /// <summary>
+    /// Calculates the width and height of a collision mask from the given margin/bounding box.
+    /// This method is used to calculate collision mask dimensions in GameMaker 2024.6 and above.
+    /// </summary>
+    public static (int Width, int Height) CalculateBboxMaskDimensions(int marginRight, int marginLeft, int marginBottom, int marginTop)
+    {
+        return (marginRight - marginLeft + 1, marginBottom - marginTop + 1);
+    }
+
+    /// <summary>
+    /// Calculates the width and height of a collision mask from a given sprite's full width and height.
+    /// This method is used to calculate collision mask dimensions prior to GameMaker 2024.6.
+    /// </summary>
+    /// <remarks>
+    /// This simply returns the width and height supplied, but is intended for clarity in the code.
+    /// </remarks>
+    public static (int Width, int Height) CalculateFullMaskDimensions(int width, int height)
+    {
+        return (width, height);
+    }
+
     private void ReadMaskData(UndertaleReader reader)
     {
+        // Initialize mask list
         uint maskCount = reader.ReadUInt32();
-        uint len = (Width + 7) / 8 * Height;
         List<MaskEntry> newMasks = new((int)maskCount);
+
+        // Read in mask data
+        (int width, int height) = CalculateMaskDimensions(reader.undertaleData);
+        uint len = (uint)((width + 7) / 8 * height);
         uint total = 0;
         for (uint i = 0; i < maskCount; i++)
         {
-            newMasks.Add(new MaskEntry(reader.ReadBytes((int)len)));
+            newMasks.Add(new MaskEntry(reader.ReadBytes((int)len), width, height));
             total += len;
         }
 
-        CollisionMasks = new(newMasks);
-
-        while (total % 4 != 0)
+        while ((total % 4) != 0)
         {
             if (reader.ReadByte() != 0)
+            {
                 throw new IOException("Mask padding");
+            }
             total++;
         }
-        Util.DebugUtil.Assert(total == CalculateMaskDataSize(Width, Height, maskCount));
+        if (total != CalculateMaskDataSize(width, height, maskCount))
+        {
+            throw new IOException("Mask data size incorrect");
+        }
+
+        // Assign masks to sprite
+        CollisionMasks = new(newMasks);
     }
-    private static void SkipMaskData(UndertaleReader reader, uint width, uint height)
+
+    private static void SkipMaskData(UndertaleReader reader, int width, int height, int marginRight, int marginLeft, int marginBottom, int marginTop)
     {
         uint maskCount = reader.ReadUInt32();
-        uint len = (width + 7) / 8 * height;
+        if (reader.undertaleData.IsVersionAtLeast(2024, 6))
+        {
+            (width, height) = CalculateBboxMaskDimensions(marginRight, marginLeft, marginBottom, marginTop);
+        }
+        else
+        {
+            (width, height) = CalculateFullMaskDimensions(width, height);
+        }
+        uint len = (uint)((width + 7) / 8 * height);
 
         uint total = 0;
         for (uint i = 0; i < maskCount; i++)
@@ -815,10 +930,10 @@ public class UndertaleSprite : UndertaleNamedResource, PrePaddedObject, INotifyP
         reader.Position += skipSize;
     }
 
-    public uint CalculateMaskDataSize(uint width, uint height, uint maskcount)
+    private uint CalculateMaskDataSize(int width, int height, uint maskCount)
     {
-        uint roundedWidth = (width + 7) / 8 * 8; // round to multiple of 8
-        uint dataBits = roundedWidth * height * maskcount;
+        uint roundedWidth = (uint)((width + 7) / 8 * 8); // round to multiple of 8
+        uint dataBits = (uint)(roundedWidth * height * maskCount);
         uint dataBytes = ((dataBits + 31) / 32 * 32) / 8; // round to multiple of 4 bytes
         return dataBytes;
     }
