@@ -37,7 +37,7 @@ namespace UndertaleModLib.Models;
 ///     #5 leads here:
 ///     SimpleList&lt;int&gt; of tileset IDs the group has
 /// </code>
-/// <see href="https://github.com/krzys-h/UndertaleModTool/wiki/Bytecode-17#tgin-new-chunk"/>.</remarks>
+/// <see href="https://github.com/UnderminersTeam/UndertaleModTool/wiki/Bytecode-17#tgin-new-chunk"/>.</remarks>
 [PropertyChanged.AddINotifyPropertyChangedInterface]
 public class UndertaleTextureGroupInfo : UndertaleNamedResource, IDisposable
 {
@@ -133,14 +133,14 @@ public class UndertaleTextureGroupInfo : UndertaleNamedResource, IDisposable
 
         writer.WriteUndertaleObjectPointer(TexturePages);
         writer.WriteUndertaleObjectPointer(Sprites);
-        if (!writer.undertaleData.IsVersionAtLeast(2023, 1))
+        if (!writer.undertaleData.IsNonLTSVersionAtLeast(2023, 1))
             writer.WriteUndertaleObjectPointer(SpineSprites);
         writer.WriteUndertaleObjectPointer(Fonts);
         writer.WriteUndertaleObjectPointer(Tilesets);
 
         writer.WriteUndertaleObject(TexturePages);
         writer.WriteUndertaleObject(Sprites);
-        if (!writer.undertaleData.IsVersionAtLeast(2023, 1))
+        if (!writer.undertaleData.IsNonLTSVersionAtLeast(2023, 1))
             writer.WriteUndertaleObject(SpineSprites);
         writer.WriteUndertaleObject(Fonts);
         writer.WriteUndertaleObject(Tilesets);
@@ -161,7 +161,7 @@ public class UndertaleTextureGroupInfo : UndertaleNamedResource, IDisposable
         // Read the pointers
         TexturePages = reader.ReadUndertaleObjectPointer<UndertaleSimpleResourcesList<UndertaleEmbeddedTexture, UndertaleChunkTXTR>>();
         Sprites = reader.ReadUndertaleObjectPointer<UndertaleSimpleResourcesList<UndertaleSprite, UndertaleChunkSPRT>>();
-        if (!reader.undertaleData.IsVersionAtLeast(2023, 1))
+        if (!reader.undertaleData.IsNonLTSVersionAtLeast(2023, 1))
             SpineSprites = reader.ReadUndertaleObjectPointer<UndertaleSimpleResourcesList<UndertaleSprite, UndertaleChunkSPRT>>();
         Fonts = reader.ReadUndertaleObjectPointer<UndertaleSimpleResourcesList<UndertaleFont, UndertaleChunkFONT>>();
         Tilesets = reader.ReadUndertaleObjectPointer<UndertaleSimpleResourcesList<UndertaleBackground, UndertaleChunkBGND>>();
@@ -169,7 +169,7 @@ public class UndertaleTextureGroupInfo : UndertaleNamedResource, IDisposable
         // Read the objects, throwing an error if the pointers are invalid
         reader.ReadUndertaleObject(TexturePages);
         reader.ReadUndertaleObject(Sprites);
-        if (!reader.undertaleData.IsVersionAtLeast(2023, 1)) 
+        if (!reader.undertaleData.IsNonLTSVersionAtLeast(2023, 1)) 
             reader.ReadUndertaleObject(SpineSprites);
         reader.ReadUndertaleObject(Fonts);
         reader.ReadUndertaleObject(Tilesets);
@@ -188,7 +188,7 @@ public class UndertaleTextureGroupInfo : UndertaleNamedResource, IDisposable
         uint texPagesPtr = reader.ReadUInt32();
         uint spritesPtr = reader.ReadUInt32();
         uint spineSpritesPtr = 0;
-        if (!reader.undertaleData.IsVersionAtLeast(2023, 1))
+        if (!reader.undertaleData.IsNonLTSVersionAtLeast(2023, 1))
             spineSpritesPtr = reader.ReadUInt32();
         uint fontsPtr = reader.ReadUInt32();
         uint tilesetsPtr = reader.ReadUInt32();
@@ -201,7 +201,7 @@ public class UndertaleTextureGroupInfo : UndertaleNamedResource, IDisposable
         count += 1 + UndertaleSimpleResourcesList<UndertaleSprite, UndertaleChunkSPRT>
                      .UnserializeChildObjectCount(reader);
 
-        if (!reader.undertaleData.IsVersionAtLeast(2023, 1))
+        if (!reader.undertaleData.IsNonLTSVersionAtLeast(2023, 1))
         {
             reader.AbsPosition = spineSpritesPtr;
             count += 1 + UndertaleSimpleResourcesList<UndertaleSprite, UndertaleChunkSPRT>
