@@ -27,16 +27,17 @@ namespace UndertaleModLib.Util
 
             public int Read(byte[] buffer, int count)
             {
+                if (count < 0)
+                {
+                    throw new ArgumentOutOfRangeException(nameof(count));
+                }
+
                 int n = _length - _position;
                 if (n > count)
                     n = count;
                 if (n <= 0)
                 {
-#if DEBUG
                     throw new IOException("Reading out of chunk bounds");
-#else
-                    return 0;
-#endif
                 }
 
                 if (n <= 8)
@@ -56,11 +57,7 @@ namespace UndertaleModLib.Util
                 int n = Math.Min(_length - _position, buffer.Length);
                 if (n <= 0)
                 {
-#if DEBUG
                     throw new IOException("Reading out of chunk bounds");
-#else
-                    return 0;
-#endif
                 }
 
                 new Span<byte>(_buffer, _position, n).CopyTo(buffer);
@@ -74,11 +71,7 @@ namespace UndertaleModLib.Util
                 int newPos = _position + 1;
                 if (newPos > _length)
                 {
-#if DEBUG
                     throw new IOException("Reading out of chunk bounds");
-#else
-                    return 0;
-#endif
                 }
 
                 _position = newPos;
@@ -87,6 +80,9 @@ namespace UndertaleModLib.Util
 
             public void Write(byte[] buffer, int count)
             {
+                if (count < 0)
+                    throw new ArgumentOutOfRangeException(nameof(count));
+
                 int i = _position + count;
                 if (i < 0)
                     throw new IOException("Writing out of the chunk buffer bounds.");
@@ -186,10 +182,15 @@ namespace UndertaleModLib.Util
 
         public string ReadChars(int count)
         {
-#if DEBUG
+            if (count < 0)
+            {
+                throw new ArgumentOutOfRangeException(nameof(count));
+            }
             if (chunkBuffer.Position + count > _length)
+            {
                 throw new IOException("Reading out of chunk bounds");
-#endif
+            }
+
             if (count > 1024)
             {
                 byte[] buf = new byte[count];
@@ -209,10 +210,15 @@ namespace UndertaleModLib.Util
 
         public byte[] ReadBytes(int count)
         {
-#if DEBUG
+            if (count < 0)
+            {
+                throw new ArgumentOutOfRangeException(nameof(count));
+            }
             if (chunkBuffer.Position + count > _length)
+            {
                 throw new IOException("Reading out of chunk bounds");
-#endif
+            }
+
             byte[] val = new byte[count];
             if (count > 0)
                 chunkBuffer.Read(val, count);
@@ -273,15 +279,16 @@ namespace UndertaleModLib.Util
 
         public string ReadGMString()
         {
-#if DEBUG
             if (chunkBuffer.Position + 5 > _length)
                 throw new IOException("Reading out of chunk bounds");
-#endif
+
             int length = BinaryPrimitives.ReadInt32LittleEndian(ReadToBuffer(4));
-#if DEBUG
-            if (chunkBuffer.Position + length + 1 >= _length)
+
+            if (length < 0)
+                throw new IOException("Invalid string length");
+            if (chunkBuffer.Position + length + 1 > _length)
                 throw new IOException("Reading out of chunk bounds");
-#endif
+
             string res;
             if (length > 1024)
             {
@@ -296,18 +303,20 @@ namespace UndertaleModLib.Util
                     chunkBuffer.Read(buf);
                 res = encoding.GetString(buf);
             }
-            
-#if DEBUG
+
             if (ReadByte() != 0)
+            {
                 throw new IOException("String not null terminated!");
-#else
-            Position++;
-#endif
+            }
+
             return res;
         }
+
         public void SkipGMString()
         {
             int length = BinaryPrimitives.ReadInt32LittleEndian(ReadToBuffer(4));
+            if (length < 0)
+                throw new IOException("Invalid string length");
             Position += (uint)length + 1;
         }
 
