@@ -1,4 +1,6 @@
-﻿using ICSharpCode.AvalonEdit;
+﻿#pragma warning disable CA1416 // Validate platform compatibility
+
+using ICSharpCode.AvalonEdit;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -68,3 +70,5 @@ namespace UndertaleModTool
         }
     }
 }
+
+#pragma warning restore CA1416 // Validate platform compatibility
