@@ -1,4 +1,6 @@
-﻿using System;
+﻿#pragma warning disable CA1416 // Validate platform compatibility
+
+using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
@@ -179,3 +181,5 @@ namespace UndertaleModTool.Windows
         }
     }
 }
+
+#pragma warning restore CA1416 // Validate platform compatibility
